Allow macOS resource bundling without AUv3
AUv3 can only be built if the Xcode generator is used. In some cases
we don't care about AUv3, and we do care about not using the Xcode
generator. This is the case when you want to generate
compile-commands.json via CMAKE_EXPORT_COMPILE_COMMANDS, which doesn't
work for the Xcode generator.