Fix duplicate MouseEvents in About window
Due to the recent introduction of a global mouse listener in order
to close the About window if the user clicks outside of it, a
regression was also introduced: Whenever the user clicks a URL in
the About text, the URL is opened in the browser twice.

I decided to introduce a dedicated MouseListener implementation
rather than piggybacking on the fact that `class About` is already
a MouseListener. This seems to have automatically eliminated the
duplicate MouseEvents bug.

It's quite interesting how difficult it is to implement the
feature where the user can click outside About to close it. Most
UI frameworks only model the situation where a component wants to
know about mouse events within the components bounds, or on the
verge of it (mouse enters, mouse leaves).

I just whipped something together and went for the first thing
that worked. It doesn't feel like a great implementation. Maybe
one day I will become an expert in components that want to know
about mouse clicks outside of it.