Improve host synchronization #264
Complete overhaul of position tracking. We move away, from an
integer that tracks the current position in ticks, to a double that
tracks the current position in quarter notes. This is congruent
with how plugin hosts expose position information to plugins.

Unrelated to #264, this commit also addresses an issue with
AllSequencer, which is responsible for serializing sequencer
and related state. It was notifying UI observers in the process,
and while working on #264 I ran into a rare but logical crash. The
implemented fix is to not notify UI observers anymore.

Open issues:
1. `OldClock` is probably not necessary anymore, so we should
   look into factoring it out.
2. A bunch of stuff is living in izzyreal/vmpc-juce that should live
   in izzyreal/mpc.
3. After 2 has been fixed, we can get rid of some code duplication
   related to computing a buffer's position in quarter notes (for the
   cases where there's no host, or the host is not playing).
4. `ppqPosition` is a misnomer in the JUCE framework. A better term
   needs to be settled on, but it should be semantically equivalent
   to "musical position in quarter notes" (see ivstprocesscontext.h)
   or "position in beats" (see clap/events.h).

We leave these open for now and issues in GitHub should be raised
for these.