Fix crash in Reaper on Intel Mac #134
Not entirely sure why this issue only occurs in Reaper on an
Intel Mac, but the fix is semantically correct -- there's nothing
that prohibits a host from invoking `prepareToPlay` repeatedly
without ever invoking `processBlock`. And the desired state of
`previousHostOutputChannelIndicesToRender` when
`prepareToPlay` returns is that it has a single dummy element, in
order to ensure proper cleaning of output buffers in the first
invocation of `processBlock` after `prepareToPlay` returns.

My original assumption that a `prepareToPlay` invocation is
always followed by a `processBlock` invocation was wrong, so in
this fix we clear `previousHostOutputChannelIndicesToRender`
before inserting the one and only element we want in this
collection at that point.