Rename ExternalClock to Clock
The original idea behind ExternalClock was that it would only
consume values produced by a plugin host. Since recently this
class also consumes values produced by the standalone host, so
rather than an external clock it is just a clock.

This prepares for moving the pulse generation code, which
currently resides in various places that consume Clock, into this
class.